Factor out metadata & posting out of parser.rs.
